agregando estados
