+++ ext/lib/site/home-presupuesto/filters-navbar/styles.styl
@@ -26,7 +26,7 @@
       padding-left: 8px
   
     .filter-dropdown
-      display: none
+      //display: none
       cursor: auto
       position: absolute
       width: 250px
